+++ src/ABLUnitRun.ts
@@ -1,9 +1,9 @@
 import { CancellationError, CancellationToken, Disposable, FileSystemWatcher, TestRun, Uri, workspace } from 'vscode'
 import { ABLResults } from './ABLResults'
-import { deleteFile, isRelativePath } from './ABLUnitCommon'
-import { ExecException, ExecOptions, exec } from 'child_process'
+import { deleteFile, Duration } from './ABLUnitCommon'
+import { SendHandle, Serializable, SpawnOptions, spawn } from 'child_process'
 import { log } from './ChannelLogger'
-import { processUpdates, updateParserInit } from 'parse/UpdateParser'
+import { processUpdates, setTimeoutTestStatus, updateParserInit } from 'parse/UpdateParser'
 
 export enum RunStatus {
 	None = 10,
@@ -17,6 +17,9 @@ export enum RunStatus {
 	Complete = 80,
 	Cancelled = 81,
 	Error = 82,
+	Killed = 83,
+	Timeout = 84,
+	Unknown = 99,
 }
 export enum RunStatusString {
 	'None' = 10,
@@ -35,11 +38,32 @@ export enum RunStatusString {
 export class ABLUnitRuntimeError extends Error {
 	constructor (message: string, public promsgError: string, public cmd?: string) {
 		super(message)
+		this.name = 'ABLUnitRuntimeError'
+	}
+}
+
+// eslint-disable-next-line @typescript-eslint/naming-convention
+export interface ITimeoutError extends Error {
+	duration: Duration
+	limit: number
+	cmd?: string
+}
+
+export class TimeoutError extends Error implements ITimeoutError {
+	duration: Duration
+	limit: number
+	cmd?: string
+
+	constructor (message: string, duration: Duration, limit: number, cmd: string) {
+		super(message)
+		this.name = 'TimeoutError'
+		this.duration = duration
+		this.limit = limit
+		this.cmd = cmd
 	}
 }
 
 export const ablunitRun = async (options: TestRun, res: ABLResults, cancellation: CancellationToken) => {
-	const start = Date.now()
 	const abort = new AbortController()
 	const { signal } = abort
 	let watcherDispose: Disposable | undefined = undefined
@@ -88,7 +112,7 @@ export const ablunitRun = async (options: TestRun, res: ABLResults, cancellation
 			throw new Error('command does not contain \'${testlist}\'')
 		}
 		cmd = cmd.replace(/\$\{testlist\}/, testlist)
-		cmd = cmd.replace(/\$\{tempDir\}/, workspace.asRelativePath(res.cfg.ablunitConfig.tempDirUri, false))
+		cmd = cmd.replace(/\$\{tempDir\}/, res.cfg.ablunitConfig.tempDirUri.fsPath)
 		const cmdSanitized = cmd.split(' ')
 
 		log.info('ABLUnit Command: ' + cmdSanitized.join(' '))
@@ -102,11 +126,11 @@ export const ablunitRun = async (options: TestRun, res: ABLResults, cancellation
 
 		const executable = res.dlc!.uri.fsPath.replace(/\\/g, '/') + '/bin/' + res.cfg.ablunitConfig.command.executable
 
-		let cmd = [ executable, '-b', '-p', res.wrapperUri.fsPath.replace(/\\/g, '/') ]
+		const cmd = [ executable, '-b', '-p', res.wrapperUri.fsPath.replace(/\\/g, '/') ]
 
 		if (process.platform === 'win32') {
 			if (res.cfg.ablunitConfig.progressIniUri) {
-				cmd.push('-basekey', 'INI', '-ininame', workspace.asRelativePath(res.cfg.ablunitConfig.progressIniUri.fsPath, false))
+				cmd.push('-basekey', 'INI', '-ininame', res.cfg.ablunitConfig.progressIniUri.fsPath)
 			}
 		} else if (process.platform === 'linux') {
 			process.env['PROPATH'] = res.propath!.toString().replace(/\$\{DLC\}/g, res.dlc!.uri.fsPath.replace(/\\/g, '/'))
@@ -114,24 +138,20 @@ export const ablunitRun = async (options: TestRun, res: ABLResults, cancellation
 			throw new Error('unsupported platform: ' + process.platform)
 		}
 
-		let tempPath = workspace.asRelativePath(res.cfg.ablunitConfig.tempDirUri, false)
-		if (isRelativePath(tempPath)) {
-			tempPath = './' + tempPath
-		}
-		cmd.push('-T', tempPath)
+		cmd.push('-T', res.cfg.ablunitConfig.tempDirUri.fsPath)
 
-		if (res.cfg.ablunitConfig.dbConnPfUri && res.cfg.ablunitConfig.dbConns.length > 0) {
-			cmd.push('-pf', workspace.asRelativePath(res.cfg.ablunitConfig.dbConnPfUri.fsPath, false))
+		if (res.cfg.ablunitConfig.dbConnPfUri && res.cfg.ablunitConfig.dbConns && res.cfg.ablunitConfig.dbConns.length > 0) {
+			cmd.push('-pf', res.cfg.ablunitConfig.dbConnPfUri.fsPath)
 		}
 
 		if (res.cfg.ablunitConfig.profiler.enabled) {
-			cmd.push('-profile', workspace.asRelativePath(res.cfg.ablunitConfig.profOptsUri, false))
+			cmd.push('-profile', res.cfg.ablunitConfig.profOptsUri.fsPath)
 		}
 
 		const cmdSanitized: string[] = []
-		cmd = cmd.concat(res.cfg.ablunitConfig.command.additionalArgs)
+		cmd.push(...res.cfg.ablunitConfig.command.additionalArgs)
 
-		let params = 'CFG=' + workspace.asRelativePath(res.cfg.ablunitConfig.config_uri.fsPath, false) + '='
+		let params = 'CFG=' + res.cfg.ablunitConfig.config_uri.fsPath + '='
 		if (res.cfg.ablunitConfig.dbAliases.length > 0) {
 			params = params + ' ALIASES=' + res.cfg.ablunitConfig.dbAliases.join(';')
 		}
@@ -173,21 +193,8 @@ export const ablunitRun = async (options: TestRun, res: ABLResults, cancellation
 		args.shift()
 
 		return new Promise<string>((resolve, reject) => {
-			res.setStatus(RunStatus.Executing)
-
+			res.setStatus(RunStatus.Running)
 			const runenv = getEnvVars(res.dlc!.uri)
-
-			const execOpts: ExecOptions = {
-				env: runenv,
-				cwd: res.cfg.ablunitConfig.workspaceFolder.uri.fsPath,
-				signal: signal,
-				timeout: 120000
-			}
-
-			const execCommand = (cmd + ' ' + args.join(' ')).replace(/\$\{DLC\}/g, res.dlc!.uri.fsPath.replace(/\\/g, '/')) + ' 2>&1'
-
-			log.info('command=\'' + cmd + ' ' + args.join(' ') + '\'\r\n', options)
-			log.info('----- ABLUnit Command Execution Started -----', options)
 
 			const updateUri = res.cfg.ablunitConfig.optionsUri.updateUri
 			if (updateUri) {
@@ -195,62 +202,123 @@ export const ablunitRun = async (options: TestRun, res: ABLResults, cancellation
 				updateParserInit()
 				log.info('watching test run update/event file: ' + updateUri.fsPath)
 				watcherUpdate = workspace.createFileSystemWatcher(updateUri.fsPath)
-				watcherDispose = watcherUpdate.onDidChange(uri => { return processUpdates(options, res.tests, updateUri) })
+				watcherDispose = watcherUpdate.onDidChange((uri) => { return processUpdates(options, res.tests, uri) })
 			}
 
-			exec(execCommand, execOpts, (err: ExecException | null, stdout: string, stderr: string) => {
-				const duration = Date.now() - start
+			const spawnOpts: SpawnOptions = {
+				signal: signal,
+				// killSignal: 'SIGKILL', // DEFAULT
+				// killSignal: 'SIGABRT', // does not actually kill the process
+				// killSignal: 'SIGINT', // works in windows
+				// killSignal: 'SIGHUP', // works in linux
+				killSignal: 'SIGTERM',
+				timeout: res.cfg.ablunitConfig.timeout,
+				env: runenv,
+				cwd: res.cfg.ablunitConfig.workspaceFolder.uri.fsPath,
+			}
 
-				if (watcherUpdate) {
-					watcherUpdate.dispose()
-				}
-				if (watcherDispose) {
-					watcherDispose.dispose()
-				}
+			log.info('command=\'' + cmd + ' ' + JSON.stringify(args) + '\'\r\n', options)
+			const testRunDuration = new Duration('TestRun')
+			const process = spawn(cmd, args, spawnOpts)
 
-				if (err) {
-					const errStr = '[err]\r\n' + JSON.stringify(err, null, 2) + '\r\n[\\err]'
-					log.error(errStr, options)
-				}
-				if (stdout) {
-					stdout = '[stdout] ' + stdout + '[\\stdout]'
-					log.info(stdout, options)
-				}
-				if (stderr) {
-					stderr = '[stderr] ' + stderr.replace(/\n/g, '\n[stderr] ')
-					log.error(stderr, options)
-				}
+			let lastError: string | undefined = undefined
 
-				if (err) {
-					const promsgError = parseRuntimeError(stdout.replace(/\r/g, ''))
-					log.debug('promsgError=' + promsgError, options)
-					if (promsgError) {
-						log.error('ABLUnit Runtime Error (code=' + err.code + '):' + promsgError, options)
-						reject(new ABLUnitRuntimeError ('ABLUnit Runtime Error (code=' + err.code + ') !', promsgError, err.cmd))
-					} else {
-						log.error('Runtime error: ' + err.message + ' (code=' + err.code + ')', options)
-						reject(err)
-					}
+			process.stderr?.on('data', (data: Buffer) => {
+				void processUpdates(options, res.tests, updateUri)
+				log.error('\t\t[stderr] ' + data.toString().trim().replace(/\n/g, '\n\t\t[stderr]'), options)
+				lastError = data.toString().trim()
+			})
+			process.stdout?.on('data', (data: Buffer) => {
+				void processUpdates(options, res.tests, updateUri)
+				log.info('\t\t[stdout] ' + data.toString().trim().replace(/\n/g, '\n\t\t[stdout]'), options)
+				lastError = undefined
+			})
+			process.once('spawn', () => {
+				res.setStatus(RunStatus.Executing)
+				log.info('----- ABLUnit Test Run Started -----', options)
+			}).on('error', (e: Error) => {
+				log.debug('process.error e=' + e)
+				res.setStatus(RunStatus.Error, 'e=' + e)
+				log.error('----- ABLUnit Test Run Error ----- (e=' + e + ')', options)
+				if (e instanceof Error) {
+					reject(e)
+				}
+			}).on('exit', (code: number | null, signal: NodeJS.Signals | null) => {
+				testRunDuration.stop()
+				log.debug('process.exit code=' + code + '; signal=' + signal + '; process.exitCode=' + process.exitCode + '; process.signalCode=' + process.signalCode + '; killed=' + process.killed)
+				void processUpdates(options, res.tests, updateUri)
+				if (process.killed || signal) {
+					setTimeoutTestStatus(options, res.cfg.ablunitConfig.timeout)
+					res.setStatus(RunStatus.Timeout, 'signal=' + signal)
+					log.info('----- ABLUnit Test Run Timeout - ' + res.cfg.ablunitConfig.timeout + 'ms ----- ' + testRunDuration, options)
+					reject(new TimeoutError('ABLUnit process timeout', testRunDuration, res.cfg.ablunitConfig.timeout, cmd))
 					return
 				}
-				if(stderr) {
-					log.error('ABLUnit Command Execution Failed (reject-3) - duration: ' + duration + '\n' + stderr, options)
-					reject(new Error ('ABLUnit Command Execution Failed (reject-3) - duration: ' + duration + '\n' + stderr))
+				if (process.killed) {
+					res.setStatus(RunStatus.Killed, 'signal=' + signal)
+					log.info('----- ABLUnit Test Run Killed - (signal=' + signal + ') ----- ' + testRunDuration, options)
+					reject(new ABLUnitRuntimeError('ABLUnit process killed', 'exit_code=' + code + '; signal=' + signal, cmd))
+					return
 				}
 
-				resolve('ABLUnit Command Execution Completed - duration: ' + duration)
+				if (code && code != 0) {
+					res.setStatus(RunStatus.Error, 'exit_code=' + code)
+					log.info('----- ABLUnit Test Run Failed (exit_code=' + code + ') ----- ' + testRunDuration, options)
+					reject(new ABLUnitRuntimeError('ABLUnit exit_code= ' + code, 'ABLUnit exit_code= ' + code + '; signal=' + signal, cmd))
+					return
+				}
+
+				res.setStatus(RunStatus.Complete, 'success')
+				log.info('----- ABLUnit Test Run Complete ----- ' + testRunDuration, options)
+				resolve('success')
+			}).on('close', (code: number | null, signal: NodeJS.Signals | null) => {
+				log.debug('process.close code=' + code + '; signal=' + signal + '; process.exitCode=' + process.exitCode + '; process.signalCode=' + process.signalCode + '; killed=' + process.killed)
+
+				watcherDispose?.dispose()
+				watcherUpdate?.dispose()
+				void processUpdates(options, res.tests, updateUri)
+
+				log.info('process.close event')
+				log.info('\tcode=' + code)
+				log.info('\tsignal=' + signal)
+				log.info('\tprocess.exitCode=' + process.exitCode)
+				log.info('\tprocess.signalCode=' + process.signalCode)
+				log.info('\tprocess.killed=' + process.killed)
+				log.info('\tprocess.pid=' + process.pid)
+				log.info('\tprocess.connected' + process.connected)
+				log.info('\tprocess.channel=' + process.channel)
+
+				if (lastError && res.status < RunStatus.Complete) {
+					if (lastError) {
+						res.setStatus(RunStatus.Error)
+						log.info('----- ABLUnit Test Run Runtime Error (exit_code=' + code + ') ----- ' + testRunDuration, options)
+						reject(new ABLUnitRuntimeError('ABLUnit runtime error', lastError, cmd))
+						return
+					}
+				}
+				if (res.status < RunStatus.Complete) {
+					const currentStatus = res.status
+					res.setStatus(RunStatus.Unknown, 'ABLUnit process closed unexpectedly, should have called exit already (RunStatus=' + currentStatus + ')')
+					reject(new ABLUnitRuntimeError('ABLUnit process closed unexpectedly', 'RunStatus=' + currentStatus + '; code=' + code + '; signal=' + signal, cmd))
+					return
+				}
+				reject(new Error('Unknown error - ABLUnit process closed but was not processed correctly'))
+			}).on('message', (m: Serializable, h: SendHandle) => {
+				// eslint-disable-next-line @typescript-eslint/no-base-to-string
+				log.info('process.on.message m=' + m.toString())
 			})
 		})
 	}
 
 	return runCommand()
-		.then(() => { return processUpdates(options, res.tests, res.cfg.ablunitConfig.optionsUri.updateUri) })
 		.then(() => {
-			log.info('----- ABLUnit Command Execution Completed -----', options)
 			return res.parseOutput(options)
-		}, (err: unknown) => {
-			log.debug('runCommand() error=' + JSON.stringify(err, null, 2), options)
-			throw err
+		}, (e: unknown) => {
+			log.debug('runCommand() error=' + JSON.stringify(e, null, 2), options)
+			if (e instanceof Error) {
+				res.thrownError = e
+			}
+			throw e
 		})
 }