+++ src/parse/TestProfileParser.ts
@@ -174,7 +174,8 @@ export class RunConfig extends DefaultRunProfile {
 	constructor (private readonly profile: IRunProfile, public workspaceFolder: WorkspaceFolder) {
 		super()
 		this.tempDirUri = this.getUri(this.profile.tempDir)
-		log.debug('tempDirUri=' + this.tempDirUri.fsPath)
+		this.timeout = this.profile.timeout
+		log.debug('tempDirUri="' + this.tempDirUri.fsPath + '"')
 		this.config_uri = Uri.joinPath(this.tempDirUri, 'ablunit.json')
 		this.profOptsUri = Uri.joinPath(this.tempDirUri, 'profile.options')
 		this.dbConnPfUri = Uri.joinPath(this.tempDirUri, 'dbconn.pf')
@@ -196,7 +197,7 @@ export class RunConfig extends DefaultRunProfile {
 		if (this.options.output.writeJson) {
 			this.optionsUri.jsonUri = Uri.joinPath(this.optionsUri.locationUri, tmpFilename.replace(/\.xml$/, '') + '.json')
 		}
-		log.debug('this.optionsUri.jsonUri=' + this.optionsUri.jsonUri?.fsPath)
+		log.debug('this.optionsUri.jsonUri="' + this.optionsUri.jsonUri?.fsPath + '"')
 		if (this.options.output.updateFile) {
 			this.optionsUri.updateUri = Uri.joinPath(this.optionsUri.locationUri, this.options.output.updateFile)
 		} else {
@@ -211,10 +212,11 @@ export class RunConfig extends DefaultRunProfile {
 			this.progressIniUri = undefined
 		}
 
-		const extraParameters = getExtraParameters(this.workspaceFolder.uri, this.profile.openedgeProjectProfile)
-		log.info('extraParameters=' + extraParameters)
+		const extraParameters = getExtraParameters(this.workspaceFolder.uri, this.profile.openedgeProjectProfile)?.split(' ')
+		// TODO - re-join quoted strings
+		log.info('extraParameters=' + JSON.stringify(extraParameters))
 		if (extraParameters) {
-			this.command.additionalArgs.push(extraParameters)
+			this.command.additionalArgs.push(...extraParameters)
 		}
 
 		const charset = getProfileCharset(this.workspaceFolder.uri, this.profile.openedgeProjectProfile)